Split CallNode into a couple files
