+++ lib/opal/nodes.rb
@@ -6,7 +6,11 @@ require 'opal/nodes/literal'
 require 'opal/nodes/variables'
 require 'opal/nodes/constants'
 require 'opal/nodes/call'
-require 'opal/nodes/call_special'
+require 'opal/nodes/call/eval'
+require 'opal/nodes/call/match3'
+require 'opal/nodes/call/js_property'
+require 'opal/nodes/call/reflection'
+require 'opal/nodes/call/require'
 require 'opal/nodes/module'
 require 'opal/nodes/class'
 require 'opal/nodes/singleton_class'
